refactor(platform specifics): modified path for android/ios specific code
