+++ src/common/mobilePlatforms/androidInterface.js
